various minor fixes

+++ duality-client/src/duality/TransferFunction.cpp
@@ -2,7 +2,7 @@
 
 TransferFunction duality::defaultTransferFunction() {
     TransferFunction tf;
-    for (uint8_t i = 0; i < 256; ++i) {
+    for (uint8_t i = 0; i < static_cast<uint8_t>(256); ++i) {
         tf[i] = {i, i, i, (i < 50) ? static_cast<uint8_t>(0) : i};
     }
     return tf;
